Code to automate step 3 of the project
Processes the data from the excel file to retrieve the correct longitude and latitude values, depending on the flag. Then constructs the url by inputting the coordinates and api key to the full url and requests the json file from SolarAPI. The 3 pieces of data are parsed from the json file returned and stored in an excel file.